Update Bounty Page
