+++ ico-site/src/app/about/about.component.html
@@ -7,7 +7,7 @@
   </div>
   <div class="screens">
   	<div class="tablet"></div>
-	<div class="poster-main abt-carousel hidden">
+	<div class="poster-main abt-carousel">
 		<div class="poster-btn poster-prev-btn"><img src="../assets/img/carousel/left.png"></div>
 		<ul class="poster-list">
 			<li class="poster-item"><a href="javascript:;"><img src="../assets/img/carousel/01_Login.jpg" width="100%" ></a></li>
