+++ ico-site/src/app/bounty/bounty.component.css
@@ -17,18 +17,34 @@ a{
 .f-blue{
   color:#a1c6e7;
 }
-.btn-blue{
-  background-color: #00a3dc;
+.btn-box{
+  margin: 0 auto;
+  text-align: center;
+  padding: 5em 0 5em 0em;
+}
+.btn-box .btn{
+  min-width: 20%;
+  margin-right: 2em;
+  margin-top: 1em;
+  margin-bottom: 1em;
+}
+.btn{
   border-radius: 3px;
   padding:15px 10px;
   min-width: 30%;
   margin: 0 auto;
   color: #fff;
-  font-size: 32px;
+  font-size: 24px;
   display: inline-block;
   box-shadow: 0px 3px 15px rgba(0,0,0,0.2);
 }
-.btn-blue:hover{
+.btn-blue{
+  background-color: #00a3dc;
+}
+.btn-blue-2{
+  background-color: #194b77;
+}
+.btn:hover{
   cursor: pointer;
   transform: translateY(2px);
   box-shadow: 0px 3px 15px rgba(0,0,0,0.4);
@@ -67,7 +83,7 @@ a{
 }
 .intor {
   text-align: center;
-  padding: 0 10% 10%;
+  padding: 0 10%;
   color:#fff;
   font-size: 14px;
 }
@@ -279,11 +295,25 @@ a{
 .pc{
   width: 100%;
 }
+.classes{
+  font-family: arial,Helvetica,sans-serif;
+  color: white;
+  text-align: center;
+  padding: 1em;
+  border-top: 1px solid white;
+  border-bottom: 1px solid white;
+  margin: 1em 0 1em 0;
+}
+.classes ul{
+  columns: 2;
+  -webkit-columns: 2;
+  -moz-columns: 2;
+}
 @media only screen and (max-width: 760px) {
   .h2 {
     width: 50%;
   }
-  .btn-blue{
+  .btn{
     font-size: 14px;
     padding:10px 15px;
   }
