+++ ico-site/src/app/bounty/bounty.component.html
@@ -3,8 +3,13 @@
   <div class="intor">
     <p>Battle between the Alliance and Horde have erupted, and a new war has begun.</p>
     <p>Summoning all champions, heroes, warlords, wizards and knights out there to support our faction - and join Morpheus League. </p>
-    <p>You will be knowns as legend, leading to explore and conquer new territories against the horde and eventually earn a place in Olympus. </p>
-    <p>This is an exclusive approved group of the most heroic, loyal and honorable supporters of Morpheuslabs.io </p>
+    <p>You will be knowns as legend, leading to explore and conquer new territories against the horde and eventually earn a place in Olympus, and total <strong>USD 300,000</strong> worth of token can be earned from this campaign.</p>
+    <br>
+    <p>This is an exclusive approved group of the most heroic, loyal and honorable supporters of Morpheus Labs</p>
+  </div>
+  <div class="btn-box">
+    <a class="btn btn-blue" href="https://goo.gl/forms/nV3wuDE75i0QZLZq2" target="_blank">Join our league now! </a>
+    <a class="btn btn-blue-2" href="https://goo.gl/forms/nV3wuDE75i0QZLZq2" target="_blank">Submit Proof of Honor</a>
   </div>
   <div class="info">
     <h2 class="h2">
@@ -15,6 +20,20 @@
     <div class="info-list">
       <img src="../../assets/img/bounty/frame_bg_top.png" alt="" class="pc">
       <div class="list-box">
+        <div class="classes">
+          <h4>Choose from one of the classes and lead Human Race to conquer unexplored territories</h4>
+          <ul>
+            <li>Knight</li> 
+            <li>Paladin</li>            
+            <li>Warlock</li> 
+            <li>Mage</li>
+            <li>Warrior</li> 
+            <li>Priest</li>
+            <li>Hunter</li>
+            <li>Templar</li> 
+          </ul>
+        </div>
+
         <ul class="list">
           <li class="li-head">
             <div class="col"> Proof of Honor <span class="mobile">&</span></div>
@@ -23,7 +42,7 @@
           <li class="li-main">
             <div class="col">
               <h3 class="h3">Telegram</h3>
-              <p><span class="f-blue">Invite your friends to join our Telegram community.</span></p>
+              <p><span class="f-blue">Invite your friends to join our <a href="http://t.me/morpheuslabs" target="_blank">Telegram</a> community.</span></p>
             </div>
             <div class="col">
               Each invite who join and stay till the end of the campaign will receive 10 MiT tokens.
@@ -31,8 +50,8 @@
           </li>
           <li class="li-main">
             <div class="col">
-              <h3 class="h3">Invite your friends to join our Telegram community.</h3>
-              <p><span class="f-blue">Share our posts from our Facebook, Linkedin & Medium page with your friends.</span></p>
+              <h3 class="h3">Facebook, LinkedIn & Medium</h3>
+              <p><span class="f-blue">Share our posts from our <a href="https://www.facebook.com/morpheusblockchain" target="_blank">Facebook</a>, <a href="https://www.linkedin.com/company/27162662/" target="_blank">LinkedIn</a> & <a href="https://medium.com/@token_34572" target="_blank">Medium</a> page with your friends.</span></p>
             </div>
             <div class="col">
               <p>Each share will earn you 1 MiT token.</p>
@@ -41,8 +60,18 @@
           </li>
           <li class="li-main">
             <div class="col"> 
+              <h3 class="h3">Twitter</h3>
+              <p><span class="f-blue">Retweet our posts on <a href="https://twitter.com/Morpheus_Labs_" target="_blank">Twitter</a>.</span></p>       
+            </div>
+            <div class="col">
+              <p>Each share will earn you 1 MiT token.</p>
+              (Min qualification: 150 followers on Twitter and account must be at least 6 months old)
+            </div>
+          </li>
+          <li class="li-main">
+            <div class="col"> 
               <h3 class="h3">Bitcointalk Signature Bounties </h3>
-              <p><span class="f-blue">Please copy from our template here</span></p>       
+              <p><span class="f-blue">Please copy from our template <a href="https://docs.google.com/document/d/1-0W_z1Rb2T1hy9FnZCv4ku51y-m6Q62YtkiLTbbfeZw/edit" target="_blank">here</a></span></p>
             </div>
             <div class="col">
               <p>Legendary/Hero = 5 MiT tokens Per Week</p>
@@ -55,7 +84,7 @@
           <li class="li-main">
             <div class="col"> 
               <h3 class="h3">Referrals</h3>
-              <p><span class="f-blue">Once you create an account on our site, you will receive an unique referral link. Be our champion by sharing our amazing project with your friends.</span></p>
+              <p><span class="f-blue">Once you create an account on our <a href="http://token.morpheuslabs.io/" target="_blank">site</a>, you will receive an unique referral link. Be our champion by sharing our amazing project with your friends.</span></p>
             </div>
             <div class="col">
               <p>You and Your friend will receive 5% bonus tokens of amount your friend purchased. </p>
@@ -109,21 +138,38 @@
               </p>
             </div>
           </li>
+          <li class="li-main">
+            <div class="col">
+              <h3 class="h3">Bug Report</h3>
+              <p><span class="f-blue">Help us spot typos or technical bugs in our website, platform etc.</span></p>
+            </div>
+            <div class="col">
+              <p>Each approved report will earn you 700 MiT tokens.
+              (Please reserve language of your choice with us before start)
+              </p>
+            </div>
+          </li>
 
           <li class="li-col-1 f-blue">
             *If there is suspicious bot activity on your account you will be banned from our bounty campaign.
           </li>
           <li class="li-col-1 f-blue">
-            Each class will have 1 chief and 3 legion leaders to help ensure our code of honours are abide in all unexplored territories. You will need to apply for these position by writing in to us @ token@morpheuslabs.io and approved accordingly based on your experience and levels.  
+            Each class will have 1 chief and 3 legion leaders to help ensure our code of honours are abide in all unexplored territories. Each class will have an guardian from Morpheus guiding you and. He or she will select chief and leaders of each class when time comes. 
           </li>
           <li class="li-col-1 f-blue">
             Code of Honour: 1. Never be aggressive & offensive in any territories, 2. For the glory of Morpheus, 3. Provide help to your legion and your league - Morpheus, 4. Summon your powerful friends to join us
+          </li>
+          <li class="li-col-1 f-blue">
+            You can contact us at token@morpheuslabs.io and depend on the question, we will try our best to answer within 24-48 hours 
           </li>
         </ul>
       </div>
       <img class="pc" src="../../assets/img/bounty/frame_bg_bot.png" alt="">
       <div class="list-bot" >
-        <a class="btn-blue" href="https://goo.gl/forms/nV3wuDE75i0QZLZq2" target="_blank">Join our league now! </a>
+        <div class="btn-box">
+          <a class="btn btn-blue" href="https://goo.gl/forms/nV3wuDE75i0QZLZq2" target="_blank">Join our league now! </a>
+          <a class="btn btn-blue-2" href="https://goo.gl/forms/nV3wuDE75i0QZLZq2" target="_blank">Submit Proof of Honor</a>
+        </div>
         <h2 class="call-action"> For Blood and honor!! </h2>
       </div>
   </div>
